+++ app/components/map.js
@@ -1,0 +1,28 @@
+/* Temporary file to from tutorial to test out environment */
+
+import React, { Component } from 'react';
+import { StyleSheet, View } from 'react-native';
+import Mapbox from '@mapbox/react-native-mapbox-gl';
+
+Mapbox.setAccessToken('pk.eyJ1IjoiYWNhLW1hcGJveCIsImEiOiJjajhkbmNjN2YwcXg0MnhzZnU2dG93NmdqIn0.jEUoPlUBoAsHAZw5GKpgiQ');
+
+export default class Map extends Component<{}> {
+  render() {
+    return (
+      <View style={styles.container}>
+        <Mapbox.MapView
+            styleURL={"mapbox://styles/aca-mapbox/cj8w8rbjnfwit2rpqudlc4msn"}
+            zoomLevel={1}
+            centerCoordinate={[-77.6109, 43.1610]}
+            style={styles.container}>
+        </Mapbox.MapView>
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
